Formatting and restructuring imports
