+++ src/basic_strategy/mode_selector.py
@@ -1,7 +1,7 @@
 import itertools
 import random
 
-from basic_strategy.hand import Card
+from src.basic_strategy.hand import Card
 
 SUITS = ["h", "c", "s", "d"]
 RANKS = ["2", "3", "4", "5", "6", "7", "8", "9", "T", "J", "Q", "K", "A"]
